Coverage for edge cases when removing directories
