+++ src/test/java/comics/logic/TestUnpack.java
@@ -1,9 +1,7 @@
 package comics.logic;
 
 import comics.commands.UnpackCommand;
-import comics.utils.BackupService;
-import comics.utils.TestUtils;
-import comics.utils.TestUtils.TestLevel;
+import comics.utils.Tools.TestLevel;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
@@ -11,12 +9,12 @@ import org.junit.jupiter.params.provider.EnumSource;
 import java.io.File;
 import java.nio.file.Files;
 
-import static comics.utils.TestUtils.TestLevel.COMMAND;
-import static comics.utils.TestUtils.TestLevel.SERVICE;
-import static comics.utils.TestUtils.copyResource;
-import static comics.utils.TestUtils.md5;
-import static comics.utils.TestUtils.runTest;
-import static comics.utils.TestUtils.today;
+import static comics.utils.Tools.TestLevel.COMMAND;
+import static comics.utils.Tools.TestLevel.SERVICE;
+import static comics.utils.Tools.copyResource;
+import static comics.utils.Tools.md5;
+import static comics.utils.Tools.runTest;
+import static comics.utils.Tools.today;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
